requirements.txt
click==8.1.3
dnspython==2.2.1
Flask==2.1.2
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
gunicorn==20.1.0
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pymongo==4.1.1
python-dotenv==0.20.0
six==1.16.0
Werkzeug==2.1.2
zipp==3.8.0
